+++ Editor/src/com/kotcrab/vis/editor/module/editor/ProjectAutoLoader.java
@@ -20,50 +20,42 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.files.FileHandle;
 import com.google.common.eventbus.Subscribe;
 import com.kotcrab.vis.editor.App;
+import com.kotcrab.vis.editor.Editor;
 import com.kotcrab.vis.editor.Log;
 import com.kotcrab.vis.editor.event.OpenSceneRequest;
 import com.kotcrab.vis.editor.event.ProjectStatusEvent;
 import com.kotcrab.vis.editor.event.ProjectStatusEvent.Status;
 import com.kotcrab.vis.editor.module.EventBusSubscriber;
 import com.kotcrab.vis.editor.util.vis.EditorException;
+import com.kotcrab.vis.editor.util.vis.LaunchConfiguration;
 
-/**
- * The purpose of this module is to dramatically speed up development by auto loading project and test scene. To start using it do the
- * following things: open VisEditor, click About from Help menu, click open app folder. Create file with name debug.this that
- * will contain two lines: first is full project path and the second one is full scene path.
- * @author Kotcrab
- */
+import java.util.Optional;
+
+/** @author Kotcrab */
 @EventBusSubscriber
-public class DevelopmentSpeedupModule extends EditorModule {
-	private static final String TAG = "DevelopmentSpeedupModule";
+public class ProjectAutoLoader extends EditorModule {
+	private static final String TAG = "ProjectAutoLoader";
 
 	private ProjectIOModule projectIO;
 
-	private FileHandle projectFile;
-	private FileHandle sceneFile;
+	private Optional<FileHandle> projectFile;
+	private Optional<FileHandle> sceneFile;
 
 	private boolean firstLoading = true;
 
 	@Override
 	public void postInit () {
-		FileHandle debugFile = new FileHandle(App.APP_FOLDER_PATH).child("debug.this");
+		LaunchConfiguration config = Editor.instance.getLaunchConfig();
+		projectFile = getAbsoluteHandleIfExists(config.projectPath);
+		sceneFile = getAbsoluteHandleIfExists(config.scenePath);
 
-		if (debugFile.exists() == false) return;
-
-		String[] lines = debugFile.readString().split("\\r?\\n");
-
-		if (lines.length < 2) {
-			Log.warn(TAG, "Debug file found but it's format is invalid, project won't be auto loaded");
+		if (projectFile.isPresent() == false) {
+			Log.warn(TAG, "Could not find project: " + config.projectPath);
 			return;
-		} else {
-			Log.debug(TAG, "Debug file found, will auto load project defined in debug file");
 		}
 
-		projectFile = Gdx.files.absolute(lines[0]);
-		sceneFile = Gdx.files.absolute(lines[1]);
-
 		try {
-			projectIO.load(projectFile);
+			projectIO.load(projectFile.get());
 		} catch (EditorException e) {
 			Log.exception(e);
 		}
@@ -74,11 +66,18 @@ public class DevelopmentSpeedupModule extends EditorModule {
 		if (firstLoading == false || projectFile == null || sceneFile == null) return;
 
 		if (event.status == Status.Loaded) {
-			if (sceneFile.exists()) {
-				App.eventBus.post(new OpenSceneRequest(sceneFile));
+			if (sceneFile.isPresent()) {
+				App.eventBus.post(new OpenSceneRequest(sceneFile.get()));
 			}
 
 			firstLoading = false;
 		}
+	}
+
+	private Optional<FileHandle> getAbsoluteHandleIfExists (String path) {
+		if (path == null) return Optional.empty();
+		FileHandle file = Gdx.files.absolute(path);
+		if (file.exists() == false) return Optional.empty();
+		return Optional.of(file);
 	}
 }