+++ Editor/CHANGES.md
@@ -2,6 +2,7 @@
 - **Added**: Scale entities tool
 - **Added**: Rotate entities tool
 - **Added**: It's now possible to set scene variables, added "Scene Variables" dialog
+- **Added**: New command line arguments: `--project` and `--scene` for auto loading project 
 - **Fixed**: Wrong entities position after opening same scene for the second time
 - **Fixed**: Sprite flip Y property was ignored after reloading scene ([#103](https://github.com/kotcrab/VisEditor/issues/103))
 - **Fixed**: Missing "Enter into Group" button after right clicking entities group ([#102](https://github.com/kotcrab/VisEditor/issues/102))